Refactor: extract ValueClassExtractor/Wrapper macros to scala-2 sources
